@@ -1,6 +1,14 @@
 name: schulcloud
 description: A Flutter–based mobile app for the HPI Schul–Cloud.
-version: 1.0.0+1
+# Version Code: MM mm pp P nn (max: 2 100 000 000)
+# MM: Major version (0 – 21)
+# mm: Minor version (0 – 99)
+# pp: Patch version (0 – 99)
+# P: Preview (2: canary, 4: alpha, 5: beta, 8: RC)
+# nn: Preview version (0 – 999)
+#
+# e.g. 1.23.4-beta.12 ≙ 01 23 04 5 012
+version: 0.3.5
 homepage: https://github.com/schul-cloud/schulcloud-flutter
 
 environment:
@@ -13,19 +21,26 @@ dependencies:
     sdk: flutter
 
   analyzer: ^0.39.4
-  black_hole_flutter: ^0.1.0
+  black_hole_flutter: ^0.2.0
   cupertino_icons: ^0.1.2
   dartx: ^0.2.0
   datetime_picker_formfield: ^1.0.0
   file_picker: ^1.4.3+2
   flare_flutter: ^1.5.8
+  flutter_absolute_path: ^1.0.6
   flutter_cached: ^4.2.4
+  flutter_deep_linking: ^0.1.0
   # The current version of flutter_downloader, 1.2.1, causes the build to fail. Until this is fixed, we rely on an older version.
   flutter_downloader: 1.1.9
-  flutter_html: ^0.10.4
+  flutter_html:
+    git:
+      url: git://github.com/Sub6Resources/flutter_html
+      ref: b5046d9
+  flutter_inappwebview: ^2.1.0+1
   flutter_native_timezone: ^1.0.4
   flutter_svg: ^0.14.2
   flutter_villains: ^1.2.0
+  font_awesome_flutter: ^8.7.0
   get_it: 4.0.0-beta3
   grec_minimal: ^0.0.3
   hive: ^1.2.0
@@ -46,29 +61,31 @@ dependencies:
   permission_handler: ^3.3.0
   provider: ^4.0.0
   rxdart: ^0.22.3
+  share:
+    git:
+      url: https://github.com/d-silveira/flutter-share.git
+      ref: 1f8b139
   sprintf: ^4.0.2
   streaming_shared_preferences: ^1.0.1
   time_machine: ^0.9.12
   transparent_image: ^1.0.0
+  uni_links: ^0.2.0
   url_launcher: ^5.1.2
-  # Using a fork until the main plugin supports transparent backgrounds:
-  # https://github.com/flutter/flutter/issues/29300
-  # webview_flutter: ^0.3.10+4
-  webview_flutter:
-    git:
-      url: git://github.com/jaumard/plugins
-      ref: feature/webview_transparency
-      path: packages/webview_flutter
 
 dev_dependencies:
   build_resolvers: ^1.1.1
   build_runner: ^1.0.0
+  flutter_driver:
+    sdk: flutter
   hive_generator:
     git:
       url: git://github.com/hivedb/hive
       ref: f8bdb804da56d0863e03316842317b27f47d2bdf
       path: hive_generator
   intl_utils: ^1.0.2
+  screenshots: ^2.1.1
+  # We need to use a version compatible with test_api 0.2.11 required by flutter_driver.
+  test: 1.9.4
 
 flutter:
   uses-material-design: true
@@ -79,6 +96,7 @@ flutter:
     - assets/empty_states/files.flr
     - assets/icon_signOut.svg
     - assets/file_thumbnails/
+    - assets/sloth_error.svg
     - assets/theme/sc/logo/logo_with_text.svg
     - assets/theme/brb/logo/logo_with_text.svg
     - assets/theme/n21/logo/logo_with_text.svg
