@@ -8,13 +8,13 @@ import 'utils.dart';
 class AppConfig {
   const AppConfig({
     @required this.name,
-    @required this.domain,
+    @required this.host,
     @required this.title,
     @required this.primaryColor,
     @required this.secondaryColor,
     @required this.accentColor,
   })  : assert(name != null),
-        assert(domain != null),
+        assert(host != null),
         assert(title != null),
         assert(primaryColor != null),
         assert(secondaryColor != null),
@@ -28,10 +28,10 @@ class AppConfig {
   static const errorColor = Color(0xFFDC2831);
 
   final String name;
-  final String domain;
-  String get baseWebUrl => 'https://$domain';
+  final String host;
+  String get baseWebUrl => 'https://$host';
   String webUrl(String path) => '$baseWebUrl/$path';
-  String get baseApiUrl => 'https://api.$domain';
+  String get baseApiUrl => 'https://api.$host';
 
   final String title;
   final MaterialColor primaryColor;
@@ -53,6 +53,9 @@ class AppConfig {
           brightness == Brightness.light ? Colors.white : null,
       fontFamily: 'PT Sans',
       textTheme: _createTextTheme(brightness),
+      buttonTheme: ButtonThemeData(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
+      ),
       dialogTheme: DialogTheme(
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(16),
