@@ -1,7 +1,13 @@
 import 'package:flutter/material.dart';
 
+import 'top_level_screen_wrapper.dart';
+
 class TopLevelPageRoute<T> extends PageRoute<T> {
-  TopLevelPageRoute({@required this.builder}) : assert(builder != null);
+  TopLevelPageRoute({
+    @required this.builder,
+    @required RouteSettings settings,
+  })  : assert(builder != null),
+        super(settings: settings);
 
   final WidgetBuilder builder;
 
@@ -12,9 +18,13 @@ class TopLevelPageRoute<T> extends PageRoute<T> {
   String get barrierLabel => null;
 
   @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-          Animation<double> secondaryAnimation) =>
-      builder(context);
+  Widget buildPage(
+    BuildContext context,
+    Animation<double> animation,
+    Animation<double> secondaryAnimation,
+  ) {
+    return TopLevelScreenWrapper(child: builder(context));
+  }
 
   @override
   Widget buildTransitions(BuildContext context, Animation<double> animation,
