@@ -1,12 +1,13 @@
 import 'dart:convert';
 import 'dart:io' as io;
 
+import 'package:black_hole_flutter/black_hole_flutter.dart';
 import 'package:dartx/dartx_io.dart';
-import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_downloader/flutter_downloader.dart';
 import 'package:hive/hive.dart';
+import 'package:get_it/get_it.dart';
 import 'package:meta/meta.dart';
 import 'package:mime/mime.dart';
 import 'package:open_file/open_file.dart';
@@ -16,6 +17,7 @@ import 'package:schulcloud/app/app.dart';
 import 'package:time_machine/time_machine.dart';
 
 import 'data.dart';
+import 'widgets/choose_destination.dart';
 
 @immutable
 class UploadProgressUpdate {
@@ -32,19 +34,19 @@ class UploadProgressUpdate {
 }
 
 extension AssociatedLocalFile on File {
-  LocalFile get localFile => services.get<FileBloc>().localFiles.get(id.value);
+  LocalFile get localFile => services.files.localFiles.get(id.value);
   bool get isDownloaded => localFile != null;
 }
 
 @immutable
-class FileBloc {
-  const FileBloc._(this.localFiles);
+class FileService {
+  const FileService._(this.localFiles);
 
   final Box<LocalFile> localFiles;
 
-  static Future<FileBloc> create() async {
+  static Future<FileService> create() async {
     final box = await Hive.openBox<LocalFile>('localFiles');
-    return FileBloc._(box);
+    return FileService._(box);
   }
 
   Future<void> openFile(File file) async {
@@ -108,14 +110,12 @@ class FileBloc {
     await localFiles.clear();
   }
 
-  Stream<UploadProgressUpdate> uploadFile({
-    @required Id<dynamic> owner,
-    Id<File> parent,
+  Stream<UploadProgressUpdate> uploadFiles({
+    @required List<io.File> files,
+    @required FilePath path,
   }) async* {
-    assert(owner != null);
-
-    // Let the user pick files.
-    final files = await FilePicker.getMultiFile();
+    assert(files != null);
+    assert(path != null);
 
     for (var i = 0; i < files.length; i++) {
       final file = files[i];
@@ -126,18 +126,19 @@ class FileBloc {
         totalNumberOfFiles: files.length,
       );
 
-      await _uploadSingleFile(file: file, owner: owner, parent: parent);
+      await _uploadSingleFile(file: file, path: path);
     }
   }
 
   Future<void> _uploadSingleFile({
     @required io.File file,
-    @required Id<dynamic> owner,
-    Id<File> parent,
+    @required FilePath path,
   }) async {
     assert(file != null);
+    logger.i('Uploading file $file');
 
     if (!file.existsSync()) {
+      logger.e("File $file doesn't exist.");
       return;
     }
 
@@ -146,15 +147,19 @@ class FileBloc {
     final mimeType = lookupMimeType(fileName, headerBytes: fileBuffer);
 
     // Request a signed url.
+    logger.d('Requesting a signed url.');
     final signedUrlResponse =
         await services.api.post('fileStorage/signedUrl', body: {
       'filename': fileName,
       'fileType': mimeType,
-      if (parent != null) 'parent': parent,
+      if (path.parentId != null) 'parent': path.parentId,
     });
     final signedInfo = json.decode(signedUrlResponse.body);
 
     // Upload the file to the storage server.
+    logger
+      ..d('Received signed info $signedInfo.')
+      ..d('Now uploading the actual file to ${signedInfo['url']}.');
     await services.network.put(
       signedInfo['url'],
       headers: (signedInfo['header'] as Map).cast<String, String>(),
@@ -162,10 +167,11 @@ class FileBloc {
     );
 
     // Notify the api backend.
+    logger.d('Notifying the file backend.');
     await services.api.post('fileStorage', body: {
       'name': fileName,
-      if (owner is! Id<User>) ...{
-        'owner': owner.value,
+      if (path.ownerId is! Id<User>) ...{
+        'owner': path.ownerId.value,
         // TODO(marcelgarus): For now, we only support user and course owner, but there's also team.
         'refOwnerModel': 'course',
       },
@@ -174,5 +180,42 @@ class FileBloc {
       'storageFileName': signedInfo['header']['x-amz-meta-flat-name'],
       'thumbnail': signedInfo['header']['x-amz-meta-thumbnail'],
     });
+    logger.i('Done uploading the file.');
   }
+
+  Future<void> uploadFileFromLocalPath({
+    @required BuildContext context,
+    @required String localPath,
+  }) async {
+    logger.i('Letting the user choose a destination where to upload '
+        '$localPath.');
+    final file = io.File(localPath);
+
+    final destination = await context.rootNavigator.push(MaterialPageRoute(
+      fullscreenDialog: true,
+      builder: (_) => ChooseDestinationScreen(
+        title: Text(context.s.file_chooseDestination_upload),
+        fabIcon: Icon(Icons.file_upload),
+        fabLabel: Text(context.s.file_chooseDestination_upload_button),
+      ),
+    ));
+
+    if (destination != null) {
+      logger.i('Uploading to $destination.');
+      await services.snackBar.performAction(
+        action: () => uploadFiles(
+          files: [file],
+          path: destination,
+        ).forEach((_) {}),
+        loadingMessage:
+            context.s.file_uploadProgressSnackBarContent(1, file.name, 0),
+        successMessage: context.s.file_uploadCompletedSnackBar,
+        failureMessage: context.s.file_uploadFailedSnackBar,
+      );
+    }
+  }
+}
+
+extension FileServiceGetIt on GetIt {
+  FileService get files => get<FileService>();
 }
