@@ -1,10 +1,11 @@
-package com.example.schulcloud;
+package org.schulcloud.android;
 
 import android.os.Bundle;
 import io.flutter.app.FlutterActivity;
 import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.plugins.share.FlutterShareReceiverActivity;
 
-public class MainActivity extends FlutterActivity {
+public class MainActivity extends FlutterShareReceiverActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
