@@ -1,7 +1,9 @@
 import 'package:black_hole_flutter/black_hole_flutter.dart';
+import 'package:dartx/dartx.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
+import '../utils.dart';
 import 'filtering.dart';
 import 'sorting.dart';
 
@@ -11,14 +13,59 @@ typedef SortFilterChangeCallback<T> = void Function(
 
 @immutable
 class SortFilter<T> {
-  const SortFilter({
-    this.sortOptions = const {},
+  SortFilter({
+    this.sorters = const {},
+    @required this.defaultSorter,
+    this.defaultSortOrder = SortOrder.ascending,
     this.filters = const {},
-  })  : assert(sortOptions != null),
+  })  : assert(sorters != null),
+        assert(sorters.isNotEmpty),
+        assert(defaultSorter != null),
+        assert(defaultSortOrder != null),
         assert(filters != null);
 
-  final Map<String, Sorter<T>> sortOptions;
+  final Map<String, Sorter<T>> sorters;
+  final String defaultSorter;
+  final SortOrder defaultSortOrder;
   final Map<String, Filter> filters;
+
+  SortFilterSelection<T> get defaultSelection {
+    return SortFilterSelection(
+      config: this,
+      sortSelectionKey: defaultSorter,
+      sortOrder: defaultSortOrder,
+      filterSelections: {
+        for (final entry in filters.entries)
+          entry.key: entry.value.defaultSelection,
+      },
+    );
+  }
+
+  /// Parses a query string generated by the web client, e.g.
+  /// https://schul-cloud.org/homework/#?dueDateFrom=2020-03-09&dueDateTo=2020-03-27&private=true&publicSubmissions=false&sort=updatedAt&sortorder=1&teamSubmissions=true
+  SortFilterSelection<T> tryParseWebQuery(Map<String, String> query) {
+    if (query.isEmpty) {
+      return defaultSelection;
+    }
+
+    return SortFilterSelection(
+      config: this,
+      sortSelectionKey: _tryParseWebQuerySorter(query) ?? defaultSorter,
+      sortOrder: SortOrderUtils.tryParseWebQuery(query) ?? defaultSortOrder,
+      filterSelections: {
+        for (final entry in filters.entries)
+          entry.key: entry.value.tryParseWebQuerySorter(query, entry.key),
+      },
+    );
+  }
+
+  String _tryParseWebQuerySorter(Map<String, String> query) {
+    final key = query['sort'];
+    return sorters.entries
+        .firstOrNullWhere(
+            (option) => key == (option.value.webQueryKey ?? option.key))
+        ?.key;
+  }
 }
 
 @immutable
@@ -40,7 +87,7 @@ class SortFilterSelection<T> {
   final SortFilter<T> config;
 
   final String sortSelectionKey;
-  Sorter<T> get sortSelection => config.sortOptions[sortSelectionKey];
+  Sorter<T> get sortSelection => config.sorters[sortSelectionKey];
   final SortOrder sortOrder;
 
   final Map<String, dynamic> filterSelections;
@@ -146,24 +193,26 @@ class SortFilterWidget<T> extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: <Widget>[
-        _buildSortSection(),
+        _buildSortSection(context),
         for (final filterKey in config.filters.keys)
           _buildFilterSection(context, filterKey),
       ],
     );
   }
 
-  Widget _buildSortSection() {
+  Widget _buildSortSection(BuildContext context) {
+    final s = context.s;
+
     return _Section(
       title: 'Order by',
       child: ChipGroup(
         children: <Widget>[
-          for (final sortOption in config.sortOptions.entries)
+          for (final sortOption in config.sorters.entries)
             ActionChip(
               avatar: sortOption.key != selection.sortSelectionKey
                   ? null
                   : Icon(selection.sortOrder.icon),
-              label: Text(sortOption.value.title),
+              label: Text(sortOption.value.title(s)),
               onPressed: () => onSelectionChange(
                   selection.withSortSelection(sortOption.key)),
             ),
@@ -176,7 +225,7 @@ class SortFilterWidget<T> extends StatelessWidget {
     final filter = config.filters[filterKey];
 
     return _Section(
-      title: filter.title,
+      title: filter.titleGetter(context.s),
       child: filter.buildWidget(
         context,
         selection.filterSelections[filterKey],
