@@ -1,5 +1,7 @@
-export 'bloc.dart';
 export 'data.dart';
+export 'routes.dart';
+export 'service.dart';
+export 'widgets/choose_destination.dart';
 export 'widgets/file_browser.dart';
 export 'widgets/file_tile.dart';
 export 'widgets/files_screen.dart';
