@@ -7,50 +7,53 @@ import 'package:pedantic/pedantic.dart';
 import 'package:schulcloud/app/app.dart';
 import 'package:schulcloud/course/course.dart';
 
-import '../bloc.dart';
 import '../data.dart';
+import '../service.dart';
 import 'app_bar.dart';
 import 'file_tile.dart';
-import 'page_route.dart';
 import 'upload_fab.dart';
 
 class FileBrowser extends StatelessWidget {
-  FileBrowser({
-    @required this.owner,
-    this.parent,
+  const FileBrowser(
+    this.path, {
     this.isEmbedded = false,
-  })  : assert(owner != null),
-        assert(parent == null || parent.isDirectory),
+  })  : assert(path != null),
         assert(isEmbedded != null);
 
-  final Entity owner;
-  Course get ownerAsCourse => owner is Course ? owner : null;
+  FileBrowser.myFiles(
+    Id<File> parentId, {
+    bool isEmbedded = false,
+  }) : this(
+          FilePath(services.storage.userId, parentId),
+          isEmbedded: isEmbedded,
+        );
 
-  final File parent;
+  final FilePath path;
+  bool get isOwnerCourse => path.ownerId is Id<Course>;
+  bool get isOwnerMe => path.ownerId == services.storage.userId;
 
   /// Whether this widget is embedded into another screen. If true, doesn't
   /// show an app bar.
   final bool isEmbedded;
 
-  CacheController<List<File>> get _filesController {
-    // The owner is either a [User] or a [Course]. Either way, the [owner] is
-    // guaranteed to have a [files] field.
-    return parent?.files?.controller ?? (owner as dynamic).files.controller;
-  }
-
   void _openDirectory(BuildContext context, File file) {
     assert(file.isDirectory);
 
-    context.navigator.push(FileBrowserPageRoute(
-      builder: (context) => FileBrowser(owner: owner, parent: file),
-    ));
+    if (isOwnerCourse) {
+      context.navigator.pushNamed('/files/courses/${path.ownerId}/${file.id}');
+    } else if (isOwnerMe) {
+      context.navigator.pushNamed('/files/my/${file.id}');
+    } else {
+      logger.e(
+          'Unknown owner: ${path.ownerId} (type: ${path.ownerId.runtimeType}) while trying to open directory ${file.id}');
+    }
   }
 
   Future<void> _downloadFile(BuildContext context, File file) async {
     assert(file.isActualFile);
 
     try {
-      await services.get<FileBloc>().openFile(file);
+      await services.files.openFile(file);
       unawaited(services.snackBar
           .showMessage(context.s.file_fileBrowser_downloading(file.name)));
     } on PermissionNotGranted {
@@ -58,7 +61,7 @@ class FileBrowser extends StatelessWidget {
         content: Text(context.s.file_fileBrowser_download_storageAccess),
         action: SnackBarAction(
           label: context.s.file_fileBrowser_download_storageAccess_allow,
-          onPressed: services.get<FileBloc>().ensureStoragePermissionGranted,
+          onPressed: services.files.ensureStoragePermissionGranted,
         ),
       ));
     }
@@ -71,7 +74,9 @@ class FileBrowser extends StatelessWidget {
 
   Widget _buildEmbedded(BuildContext context) {
     return CachedRawBuilder<List<File>>(
-      controller: _filesController,
+      // The owner is either a [User] or a [Course]. Either way, the [owner] is
+      // guaranteed to have a [files] field.
+      controller: path.files.controller,
       builder: (context, update) {
         if (update.hasError) {
           return ErrorScreen(update.error, update.stackTrace);
@@ -91,20 +96,73 @@ class FileBrowser extends StatelessWidget {
   }
 
   Widget _buildStandalone(BuildContext context) {
+    FileBrowserAppBar buildLoadingErrorAppBar(
+      dynamic error, [
+      Color backgroundColor,
+    ]) {
+      return FileBrowserAppBar(
+        title: error?.toString() ?? context.s.general_loading,
+        backgroundColor: backgroundColor,
+      );
+    }
+
+    Widget appBar;
+    if (isOwnerCourse) {
+      appBar = CachedRawBuilder<Course>(
+        controller: path.ownerId.controller,
+        builder: (context, update) {
+          if (!update.hasData) {
+            return buildLoadingErrorAppBar(update.error);
+          }
+
+          final course = update.data;
+          if (path.parentId == null) {
+            return FileBrowserAppBar(
+              title: course.name,
+              backgroundColor: course.color,
+            );
+          }
+
+          return CachedRawBuilder<File>(
+            controller: path.parentId.controller,
+            builder: (context, update) {
+              if (!update.hasData) {
+                return buildLoadingErrorAppBar(update.error, course.color);
+              }
+
+              final parent = update.data;
+              return FileBrowserAppBar(
+                title: parent.name,
+                backgroundColor: course.color,
+              );
+            },
+          );
+        },
+      );
+    } else if (path.parentId != null) {
+      appBar = CachedRawBuilder<File>(
+        controller: path.parentId.controller,
+        builder: (context, update) {
+          if (!update.hasData) {
+            return buildLoadingErrorAppBar(update.error);
+          }
+
+          final parent = update.data;
+          return FileBrowserAppBar(title: parent.name);
+        },
+      );
+    } else if (isOwnerMe) {
+      appBar = FileBrowserAppBar(title: context.s.file_files_my);
+    }
+
     return Scaffold(
       appBar: PreferredSize(
         preferredSize: AppBar().preferredSize,
-        child: FileBrowserAppBar(
-          backgroundColor: ownerAsCourse?.color,
-          title: parent?.name ?? ownerAsCourse?.name ?? context.s.file_files_my,
-        ),
+        child: appBar,
       ),
-      floatingActionButton: UploadFab(
-        ownerId: owner.id,
-        parentId: parent?.id,
-      ),
+      floatingActionButton: UploadFab(path: path),
       body: CachedBuilder<List<File>>(
-        controller: _filesController,
+        controller: path.files.controller,
         errorBannerBuilder: (_, error, st) => ErrorBanner(error, st),
         errorScreenBuilder: (_, error, st) => ErrorScreen(error, st),
         builder: (context, files) {
