@@ -5,8 +5,8 @@ import 'package:flutter_cached/flutter_cached.dart';
 import 'package:schulcloud/app/app.dart';
 import 'package:schulcloud/course/course.dart';
 
+import '../data.dart';
 import 'file_browser.dart';
-import 'page_route.dart';
 import 'upload_fab.dart';
 
 class FilesScreen extends StatelessWidget {
@@ -14,7 +14,9 @@ class FilesScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return FancyScaffold(
       appBar: FancyAppBar(title: Text(context.s.file)),
-      floatingActionButton: UploadFab(ownerId: services.storage.userId),
+      floatingActionButton: UploadFab(
+        path: FilePath(services.storage.userId),
+      ),
       sliver: SliverList(
         delegate: SliverChildListDelegate([
           _CoursesList(),
@@ -56,16 +58,10 @@ class _CourseCard extends StatelessWidget {
 
   final Course course;
 
-  void _showCourseFiles(BuildContext context) {
-    context.navigator.push(FileBrowserPageRoute(
-      builder: (context) => FileBrowser(owner: course),
-    ));
-  }
-
   @override
   Widget build(BuildContext context) {
     return FlatMaterial(
-      onTap: () => _showCourseFiles(context),
+      onTap: () => context.navigator.pushNamed('/files/courses/${course.id}'),
       child: SizedBox(
         height: 48,
         child: Row(
@@ -86,14 +82,7 @@ class _UserFiles extends StatelessWidget {
     return FancyCard(
       title: context.s.file_files_my,
       omitHorizontalPadding: true,
-      child: CachedRawBuilder(
-        controller: services.storage.userId.controller,
-        builder: (context, update) {
-          return update.hasData
-              ? FileBrowser(owner: update.data, isEmbedded: true)
-              : Container();
-        },
-      ),
+      child: FileBrowser.myFiles(null, isEmbedded: true),
     );
   }
 }
